Support for health check before processing
Right now it is possible to attempt to send messages before
the downstream application is listening. What this happens
messages can get stuck in flight and have to be retried.

When a health check path is given then the downstream host must
repond on a 200 at the path before message processing will begin